AI Optimization with Genetic Algorithms
Optimization with Genetic Algorithms
In the traveling salesperson problem, a salesperson, who lives in one of the cities, is expected to make a round trip visiting all the other cities and returning home. (It doesn't actually matter which city is the starting point.) The requirement is that the total distance travelled be as small as possible.